fix: Add babel-plugin-styled-components library so debugging shows name not hash
Without this babel plugin when you try to debug in Dev Tools > Elements,
it displays the components with the styled-component name shown as a hash i.e.
`<div class="sc-EHOje dGUdfn">`, whereas previously with CSS/Sass it showed the class name.

This babel plugin displays the style-component name that we specified instead of a
meaningless hash i.e. `<div class="App__DivContent-sc-1l7hwf9-0 jhbGUD">`, where `DivContent`
is the name of the styled-component that we created.