feat: Relates to #337. Part 1 of 2 - Account view - Change header account design
* Change Account screen so its header uses the Information component with a big identicon, a short address,
so it is clickable, and with identicon on left of name and address.

* Apply extra bottom margin between name and account address when Information component used in the header

* Update header so it uses Flexbox according to docs as intended

* Since we're using Flexbox without placeholder spacing for the left and right header buttons, when the title
is shown in the header it is offset if there is only a left or right button. So a `titleOffset` prop
is used on pages where the Header component is used to indicate if the title needs to be offset to the left
or right so it appears centered in the screen. We also pass a `screen` prop (named after the component where the
Header component is being used so we can tailor the header spacing on specific screens.

* Re-order props alphabetically

* QUESTION - should we consider always adding placeholder spacing for where both right and left buttons in header would be as an alternative so no offset for the title is require?